Organizes randomized results from .xlsx
EthoVision outputs  behavior results of mice via an excel file. The problem however is that since the order of animals was randomized for the experiment it is very annoying to analyze in the order the experiments were conducted.
This code simply reorganizes the results by several grouping parameters and outputs them as a new .xls file, in which analysis can be easily done.
Note that all the row and column numbers and all the file locations are specific for the last file I ran through it and will require to be changed in accordance to the file being used and its organization.

Basically what we need the script to do is create two tables, each divided by cage, animal and day, containing one of two results: exploration time of the mobile object and exploration time of the immobile object.