Initialization
This script takes all the animals, by cage, number and original orientation, and generates the proper location of their test phase objects.
OPS is a test in which the animal is placed in an arena with two objects opposing each other. The objects have 4 possible starting orientations around the arena.
During the test phase one of the objects is moved two spaces to one of the sides. This script calculates the final location of the two objects for each animal and exports the lists (in a randomized order) to .csv files.