Initialization
In this test there are 8 arms, an animal enter the start arm and goes to the goal arm where it is rewarded. Shortly afterwards this is repeated with another arm (at a distance of 2 from the goal arm) is opened (distraction arm).
This script takes the animals and generates conditions for them so that they have random start location and goal arm side per day, with repetitions of the same conditions per day. The conditions are exported to .csv files.
The function create_days generates as many files with as many repetitions per day as the user wishes.